Pull and refactor annotation action
